fix equipment creation and deletion
-fix equipment creation and deletion
-update header
