+++ helpmet-client/src/components/Header.jsx
@@ -26,9 +26,6 @@ const Header = () => {
                 <Link to='/about'>
                   <li>About</li>
                 </Link>
-                <Link to='/analytics'>
-                  <li>Analytics</li>
-                </Link>
                 <Link to='/profile'>
                   {currentUser ? (
                     <img src={currentUser.profilePicture} alt="profile" className='h-7 w-7 rounded-full object-cover'/>
