+++ helpmet-client/src/pages/Analytics.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef } from 'react';
-import { MapContainer, TileLayer, Marker, Popup } from 'react-leaflet';
-import 'leaflet/dist/leaflet.css';
+// import { MapContainer, TileLayer, Marker, Popup } from 'react-leaflet';
+// import 'leaflet/dist/leaflet.css';
 import { Chart } from 'chart.js/auto'; // Importing Chart.js
 
 const Analytics = () => {
